Login
Se agrego el login, moviendo ciertos métodos del UserManager a Login
